+++ nhe/geo/surface/__init__.py
@@ -1,0 +1,4 @@
+"""
+Package :mod:`nhe.geo.surface` contains base and specific implementations
+of earthquake rupture surfaces.
+"""
