--- nhe/surface/__init__.py
@@ -1,4 +1,0 @@
-"""
-Package :mod:`nhe.surface` contains base and specific implementations
-of earthquake rupture surfaces.
-"""
