+++ tests/geo/surface/_planar_test_data.py
