+++ tests/geo/surface/planar_test.py
@@ -1,7 +1,7 @@
 import unittest
 
 from nhe.geo import Point
-from nhe.surface.planar import PlanarSurface
+from nhe.geo.surface.planar import PlanarSurface
 import _planar_test_data as test_data
